driver: move control lines into specific drivers
also move common.hpp / driver.hpp into driver folder & add cdcadm break support
